+++ manual_test.sh
@@ -4,6 +4,8 @@ echo "== TESTING -p how much wood can a woodchuck chuck if the woodchuck is 5lb 
 cargo run -- -p how much wood can a woodchuck chuck if the woodchuck is 5lb and the wood is oak && \
 echo "== TESTING -p sing me a song -s the end ==" && \
 cargo run -- -p sing me a song -s the end && \
+echo "== TESTING -p tell a great joke -t 0 ==" && \
+cargo run -- -p tell a great joke -t 0 && \
 echo "== TESTING --models ==" && \
 cargo run -- --models && \
 echo "== TESTING -m code-davinci-002 ==" && \